+++ doc/CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   exit PKS Edit in case of an internal error.
 - Document Types may be hidden now in the file selector. This will cause the extension of the corresponding document
   type not being added to the list of selectable file extensions in file dialogs (to keep this list clear).
-- Dockerfile syntax support was added.
+- Dockerfile, Julia and Swift Programming Language syntax support was added.
 
 ### Improved
 - The Align Text function is more intuitive to use now.